:sparkles: Add support for configuration via env variables
Closes #8
